Freight-add ignores common unused build files
Current behavior of freight-add treats the first unrecognized file type as the
beginning of the repositories. Common build artifacts, like .build and .changes
files are unrecognized by freight and cause freight-add to prematurely begin
collecting repositories. This can cause problems if those files are in the
directory being imported and a glob like * or package_* is used. This patch
seeks to address that by recognizing .build and .changes files, but doing
nothing with them.