Also update binary cache if package filesize has changed
This commit adds a new function to apt.sh to grab the expected binary size from
the control file for a package. It then uses that function to check if the
filesize of a package has changed or disagrees with the cached control file for
that package. If it has changed or there isn't agreement, the cached control is
regenerated.