Fail hard if -c is used with a config file that does not exist
Presently, if freight commands are invoked with -c $file and $file does
not exist, freight happily continues and loads defaults from conf.sh and
other conf files. This can be confusing. This commit updates conf.sh to
bail out if $CONF is passed in but the file that $CONF points to does
not exist.